# docs/security/identity-local-auth.openapi.json
{
  "components": {
    "schemas": {
      "AcceptedResponse": {
        "properties": {
          "status": {
            "const": "accepted",
            "default": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "AcceptedResponse",
        "type": "object"
      },
...
      "ActionProofResponse": {
        "properties": {
          "expires_in": {
            "const": 300,
            "default": 300,
            "title": "Expires In",
            "type": "integer"
          },
          "proof": {
            "minLength": 1,
            "title": "Proof",
            "type": "string"
          }
        },
        "required": [
          "proof"
        ],
        "title": "ActionProofResponse",
        "type": "object"
      },
...
      "ChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "challenge": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Challenge",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "challenge"
        ],
        "title": "ChallengeRequest",
        "type": "object"
      },
      "CompletedResponse": {
        "properties": {
          "reauthentication_required": {
            "default": true,
            "title": "Reauthentication Required",
            "type": "boolean"
          },
          "status": {
            "const": "completed",
            "default": "completed",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "CompletedResponse",
        "type": "object"
      },
      "EmailChangeCompleteRequest": {
        "additionalProperties": false,
        "properties": {
          "grant": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Grant",
            "type": "string",
            "writeOnly": true
          },
          "recent_auth_proof": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Recent Auth Proof",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "recent_auth_proof",
          "grant"
        ],
        "title": "EmailChangeCompleteRequest",
        "type": "object"
      },
      "EmailChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "new_email": {
            "format": "email",
            "title": "New Email",
            "type": "string"
          },
          "recent_auth_proof": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Recent Auth Proof",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "recent_auth_proof",
          "new_email"
        ],
        "title": "EmailChangeRequest",
        "type": "object"
      },
      "EnrollmentStartRequest": {
        "additionalProperties": false,
        "properties": {
          "grant": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Grant",
            "type": "string",
            "writeOnly": true
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "grant",
          "password"
        ],
        "title": "EnrollmentStartRequest",
        "type": "object"
      },
      "FactorEnrollmentResponse": {
        "properties": {
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recovery Codes",
            "type": "array"
          },
          "status": {
            "const": "enrolled",
            "default": "enrolled",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "recovery_codes"
        ],
        "title": "FactorEnrollmentResponse",
        "type": "object"
      },
...
      "FactorSetupResponse": {
        "properties": {
          "challenge": {
            "title": "Challenge",
            "type": "string"
          },
          "expires_in": {
            "const": 300,
            "default": 300,
            "title": "Expires In",
            "type": "integer"
          },
          "provisioning_uri": {
            "title": "Provisioning Uri",
            "type": "string"
          }
        },
        "required": [
          "challenge",
          "provisioning_uri"
        ],
        "title": "FactorSetupResponse",
        "type": "object"
      },
      "FactorVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "challenge": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Challenge",
            "type": "string",
            "writeOnly": true
          },
          "code": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Code",
            "type": "string",
            "writeOnly": true
          },
          "method": {
            "default": "totp",
            "enum": [
              "totp",
              "recovery_code"
            ],
            "title": "Method",
            "type": "string"
          }
        },
        "required": [
          "challenge",
          "code"
        ],
        "title": "FactorVerifyRequest",
        "type": "object"
      },
      "InvitationRequest": {
        "additionalProperties": false,
        "properties": {
          "department_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Department Id"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "manager_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Manager Id"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "role_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Role Id"
          }
        },
        "required": [
          "email",
          "name"
        ],
        "title": "InvitationRequest",
        "type": "object"
      },
...
      "LocalAuthChallenge": {
        "additionalProperties": false,
        "properties": {
          "challenge": {
            "minLength": 1,
            "title": "Challenge",
            "type": "string"
          },
          "expires_in": {
            "const": 300,
            "default": 300,
            "title": "Expires In",
            "type": "integer"
          },
          "status": {
            "enum": [
              "mfa_required",
              "enrollment_required"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "challenge"
        ],
        "title": "LocalAuthChallenge",
        "type": "object"
      },
...
      "PasswordChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "recent_auth_proof": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Recent Auth Proof",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "recent_auth_proof",
          "password"
        ],
        "title": "PasswordChangeRequest",
        "type": "object"
      },
...
      "RecentAuthenticationRequest": {
        "additionalProperties": false,
        "properties": {
          "factor": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Factor",
            "type": "string",
            "writeOnly": true
          },
          "intended_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Intended Email"
          },
          "intended_password": {
            "anyOf": [
              {
                "format": "password",
                "maxLength": 128,
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Intended Password"
          },
          "intended_recovery_operation": {
            "anyOf": [
              {
                "enum": [
                  "factor_recovery",
                  "credential_and_factor_recovery",
                  "verified_address_recovery"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Intended Recovery Operation"
          },
          "method": {
            "default": "totp",
            "enum": [
              "totp",
              "recovery_code"
            ],
            "title": "Method",
            "type": "string"
          },
          "operation": {
            "enum": [
              "password_change",
              "email_change",
              "factor_replace",
              "recovery_codes",
              "assisted_recovery"
            ],
            "title": "Operation",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "target_user_id": {
            "exclusiveMinimum": 0,
            "title": "Target User Id",
            "type": "integer"
          }
        },
        "required": [
          "password",
          "factor",
          "target_user_id",
          "operation"
        ],
        "title": "RecentAuthenticationRequest",
        "type": "object"
      },
...
      "ResetCompleteRequest": {
        "additionalProperties": false,
        "properties": {
          "grant": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Grant",
            "type": "string",
            "writeOnly": true
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "grant",
          "password"
        ],
        "title": "ResetCompleteRequest",
        "type": "object"
      },
      "ResetRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ResetRequest",
        "type": "object"
      },
      "TokenResponse": {
        "description": "Schema for JWT token response.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "post_login_redirect_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Post Login Redirect To"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserBrief"
          }
        },
        "required": [
          "access_token",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Native contracts: #198\u2013#200 handlers implemented for component verification; #201 reserved. Production admission remains closed.",
    "title": "RiskHub reserved native identity contract",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/local/email/change": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "bearer + recent proof",
        "x-handler-implemented": true,
        "x-implementation-issue": 200,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/email/confirm": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailChangeCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "bearer + recent proof + new-address grant",
        "x-handler-implemented": true,
        "x-implementation-issue": 200,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/enrollment/start": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalAuthChallenge"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "invitation grant",
        "x-handler-implemented": true,
        "x-implementation-issue": 198,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
...
    "/api/v1/auth/local/mfa/confirm": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorEnrollmentResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "setup challenge",
        "x-handler-implemented": true,
        "x-implementation-issue": 199,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/mfa/setup": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorSetupResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "enrollment challenge",
        "x-handler-implemented": true,
        "x-implementation-issue": 199,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/mfa/verify": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "password + bound MFA challenge",
        "x-handler-implemented": true,
        "x-implementation-issue": 199,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/password/change": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "bearer + recent proof",
        "x-handler-implemented": true,
        "x-implementation-issue": 200,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/password/reset/complete": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "reset grant",
        "x-handler-implemented": true,
        "x-implementation-issue": 200,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/password/reset/request": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "public, generic result",
        "x-handler-implemented": true,
        "x-implementation-issue": 200,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/auth/local/recent-auth": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecentAuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionProofResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "bearer + password + factor",
        "x-handler-implemented": true,
        "x-implementation-issue": 199,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
...
    "/api/v1/users/invitations": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "platform admin",
        "x-handler-implemented": true,
        "x-implementation-issue": 198,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/users/{user_id}/invitations/cancel": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "platform admin",
        "x-handler-implemented": true,
        "x-implementation-issue": 198,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
    "/api/v1/users/{user_id}/invitations/resend": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "platform admin",
        "x-handler-implemented": true,
        "x-implementation-issue": 198,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
...
    "/api/v1/users/{user_id}/password-reset": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            },
            "description": "Reserved contract; handler not released"
          }
        },
        "x-authentication": "platform admin",
        "x-handler-implemented": true,
        "x-implementation-issue": 200,
        "x-profile": "password/none (not released)",
        "x-transaction-owner": "identity/auth service workflow, never endpoint commit"
      }
    },
...
  }
}